In EventuallyPersistentStore, move stop task funcs to deinitialize() from dtor
In the class EventuallyPersistentStore, after an instance is created, some
operations are done to create tasks, etc in the func initialize(). However,
these are undone during dtor of the class. This is not always safe as
flagged by the thread sanitizer in
http://cv.jenkins.couchbase.com/job/ep-engine-threadsanitizer-master/1479/

This commit does the undoing of initialize() stuff in deinitialize() before
delete.

Change-Id: Ie090fd74b7bad77642666d85969c97ae5001a027
Reviewed-on: http://review.couchbase.org/68974
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>