Move the scope of the parent monitor
The parent monitor object was a global std::unique_ptr which
the runtime generated an atexit handler to clean up. On Windows
it seems to be a bit random which thread it use to run these
atexit handlers, and in some cases the parent monitor would
run them (and hence wait for itself to be done).

Change-Id: Ia34c0147d5d17098f6fdd7b0cd23a7c05d347907
Reviewed-on: http://review.couchbase.org/69028
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>