Synchronize the start of the threads in incr/decr test
Let each thread wait for the other thead to start before running
the loop (and reduce the number of iterations now that we've
synchronized the start).

Initially I used a condition variable for this, but it turns
out that the clang analyser didn't follow all of the lambda
function and reported the variable used for the predicate as
not read ;-) The simple solution would be to "busy-wait"
using memcached and test for key existence. Each operation
require a number of context switches (to another process)
so it shouldn't be that bad anyway..

Change-Id: I4e142c7e2280cdc9a823cb37bd2a12634e1eb9a8
Reviewed-on: http://review.couchbase.org/69063
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>