MB-20746: Delete bloom filter only if temporary filter is created
During compaction, a temporary filter is only created in the
bloom filter callback function. There is a possibility that this
function will not get invoked if there is no data present during
compaction. So, before swapping the filter, ensure that a temporary
filter is created before deleting the original filter.

Change-Id: I4fcf11d32674be8c28dbda02724b40d3d37e335a
Reviewed-on: http://review.couchbase.org/69100
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Sriram Ganesan <sriram@couchbase.com>