Remove ASSERT_NO_THROW for conn.authenticate
The "annoying" part with that macro is that it doesn't print
any information about the exception it received making it
more or less impossible to figure out what went wrong.

Just remove the macro because if the exception _is_ thrown
it is caught "higher up" in the chain where it dumps the
content of the exception.

Change-Id: I5dd12c155510d1aa04c21933e7dec44e7d95a991
Reviewed-on: http://review.couchbase.org/69137
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>