MB-21513: Race condition accessing listening port
The ListeningPort object is stored in a vector which is
(currently) being protected by the stats mutex (which we should
split up). This patch refactored the "struct listening_port"
to "class ListeningPort" and const'd some of the members which
cannot change dynamically. It protects the access to the vector
where we previously tried to access it without the lock.

Change-Id: I3cae24e6ace3a344bde75b515a1571308cc06136
Reviewed-on: http://review.couchbase.org/69326
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>