MB-21512: Don't shutdown EngineManager twice
Previously we called the shutdown manually, and later on
it was called from the destructor without checking if
it was already shut down.

In addition to that we had some cases where we could race
during shutdown by just checking an atomic variable if
we should add stuff to the internal queues or not. We
have to hold the mutex so that another thread can't change
that state right after we checked it, but before we added
our items to the list.

Change-Id: I7686e549f2e80bbdc165d608513b882774fe658e
Reviewed-on: http://review.couchbase.org/69335
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>