Make memcached core aware of the XATTR datatype
This patch adds support for XATTRs in the mutation commands
implemented by memcached: set, add, replace, append, prepend
incr, decr and the DCP mutation command.

Note that this change is all about letting the internals of
the core operate on values that might have xattr's associated
with them. No xattrs got hurt or was transferred on the wire
with this change.

Change-Id: I6ec0981a769904f239c4e49a91b2c1ae49c09c24
Reviewed-on: http://review.couchbase.org/68557
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>