Refactor: Use a context obj for get requests
This is a preparation for making the core XATTR aware where
we may strip them off before sending the data to the client.

The patch tries to simplify the error handling and the
nesting by using different internal states. As an extra
"optimization" we no longer copy the inflated buffer into
yet another allocated buffer, but rather use the buffer
the inflation method allocated since the context object
is alive until all data has been transmitted.

Change-Id: I367480a086be1ca58f1aa7fcfe5ebdf27f3a1031
Reviewed-on: http://review.couchbase.org/70058
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>