MB-21763: Collection aware engine API
Update all API methods that work with keys so that collection aware
clients can invoke collection operations and legacy clients can work
as normal.

To achieve this API we bring namespaces into the API.

Legacy clients can only operate in the default namespace.

Collections are exposed as a namespace, ep-engine will accept get/set
against this namespace, but eventually will be policing keys against
collection configuration.

We additionally expose a system namespace, a place for our own
documents. EP-engine will be using the system namespace for
interleaving the creation and deletion of collections in the users
datastream and in the future the core may find need to also create
documents in this namespace. Initially though ep-engine will deny
external requests to use that namespace, but it exists for future use.

engine_testapp is given a item_namepace flag which is specifies the
default namespace in which tests will operate.

Change-Id: Icc9a42c220ccbfd9112940db0867886d09b33111
Reviewed-on: http://review.couchbase.org/70242
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>