Return stats if dcpDcpVbTakeoverStats returns ENGINE_SUCCESS
If we could not find a stream or the stream type was not STREAM_ACTIVE
or the stream pointer could not be cast as an ActiveStream pointer or
the stream status is STREAM_DEAD the function returns ENGINE_SUCCESS
but no stats are obtained.

This patch ensures that in these cases ENGINE_KEY_ENOENT is returned.
In addition the STREAM_DEAD check is removed from
DcpProducer::addTakeoverStats as there is one in
ActiveStream::addTakeoverStats, which returns stats.

The patch also ensure that we only call getVBucket once,
(in doDcpVbTakeoverStats) and then pass through to the sub-functions.

Change-Id: I6d355ee211383b3543e9522afdc8e46994a6ccd8
Reviewed-on: http://review.couchbase.org/70295
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Poonam D <poonam@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>