MB_21787: Remove log message from couch-kvstore for non-existent key
If an arithmetic operation is performed on a key which isn't resident,
on a full-eviction bucket then the following message is often printed
in the logs:

    WARNING (default) Failed to fetch data from database, vBucket=0
    key=Key_19 error=document not found [none]

This is unnecessarily verbose - this is not actually a warning in this
case and can quickly fill up the logs.

(Note: the reason this is shown is that prior to full-eviction, all
metadata would be resident and hence it would be unexpected to attempt
to read from disk a key which doesn't exist). However this is no
longer the case).

Change-Id: I6ac4f2b1f8d43700ff6869da8a6670e21454b3a0
Reviewed-on: http://review.couchbase.org/70345
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>