Simplify MultiLockHolder
Previously MultiLockHolder stored the states of its locks in order
to facilitate the ability to `unlock` outside of RAII (much like
std::unique_lock).

Given this feature isn't actually used, this moves the `unlock`
and `lock` methods to private visibility and removes the now
unnecessary logic which tracks the lock state at runtime.

Change-Id: I8e5f6701f239f31b8e354e9f37d1f81464de5c30
Reviewed-on: http://review.couchbase.org/70403
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>