Refactor: Use cb::compress instead of snappy in subdoc
Previously we used the "dynamic buffer" in the connection object
to temporarily store a copy of the inflated object (if the object
was stored compressed in memory). There is however one problem
with this method, if there is a race condition on updating the
document, we might however keep on growing the buffer as we
don't "release" this segment back in the case where we want
to start over.

Change-Id: I77e38e3a9a7303eec2c83aa9c9b55565fff9d6b8
Reviewed-on: http://review.couchbase.org/70408
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>