Fix const qualifiers in dtrace header
After converting code to C++, memcached fails to compile
because dtrace drops const qualifiers when generating the
header from probes. As we know for sure that we don't use
'char *' in the probes, it is pretty safe to replace all
occurences with sed to fix the build.

Also update list of object files for dtrace ELF generator.

Change-Id: Ic174f43b3cedc0c834cf996987d18b8d9b90ed51
Reviewed-on: http://review.couchbase.org/70468
Tested-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>