Allow xattr modifications be performed on non-json doc
The subdocument API returns an error if the document isn't
JSON. This check used to be performed when looking up the
document to modify, but that restriction would limit the
extended attributes to only work for JSON documents.

We need to move this check to where we're going to modify
the the actual data payload so that we can perform operations
on the xattrs stored on all object types.

Change-Id: I3f318240310ed5e9db7aebb9fe06be813f264143
Reviewed-on: http://review.couchbase.org/70554
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>