Modify DISALLOW_COPY_AND_ASSIGN to delete instead of relying on visibility
By deleting the functions instead of just declaring them it becomes
harder to use the macros wrong. It also means more meaningful compiler
errors are generated if someone tries to use the deleted functions.

This also removes an unused private variable that can now be statically
proven by the compiler as unused (and therefore generates a warning).

Change-Id: I13c7a04a8e56bc62a5a6b9f765889e88428a7388
Reviewed-on: http://review.couchbase.org/70587
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>