MB-21916: StoredDocKey - A DocKey container and test code
This commit introduces StoredDocKey - a container to replace
our use of std::string for document keys. The container can be used
in std::map/set etc... and allows for us to store all items that are
needed to reference a document.

- the key (and the key size)
- the document namespace (a critical piece for Collections)

StoredDocKey will store a key of n-bytes in an n+1 std::string.
The extra byte is where we store the DocNamespace.

Change-Id: I1221e55d0081427c5f477773234ebbab482a6226
Reviewed-on: http://review.couchbase.org/70680
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: buildbot <build@couchbase.com>