MB-16337: Fix for intermittent test_access_scanner failure (followup)
The original fix for this MB (d06d7adc) made an error in setting the
alog_task_time - it set it just before the engine was reloaded, where
the config will get re-parsed again and hence the updated value is ignored.

Fix by passing the new alog_task_time as part of the config string
when reloading the engine.

Change-Id: I089a941338ea732865b020ab64ef452cb3cea5ae
Reviewed-on: http://review.couchbase.org/71313
Reviewed-by: Manu Dhundi <manu@couchbase.com>
Tested-by: buildbot <build@couchbase.com>