MB-21917: SerialisedDocKey and test code
Some parts of ep-engine store a document key as part of a larger
memory allocation, e.g. MutationLog has the key inline with other
metadata that it wishes to put to file in a single write.

SerialisedDocKey serves this purpose, it can store all data that
DocKey/StoredDocKey carries, but exists to be allocated into a
pre-allocated buffer.

StoredValue and MutationLog make use of this key object.

Change-Id: I21c8f4052b0899e753809d0cd93eb6f2cae0f963
Reviewed-on: http://review.couchbase.org/70722
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: buildbot <build@couchbase.com>