MB-21784: Don't use opencheckpointid to determine if in backfill phase
The opencheckpointid of a bucket can be zero after a rollback.
If an opencheckpointid was zero it was assumed that the vbucket was in
backfilling state.  This caused the producer stream request to be stuck
waiting for backfilling to complete.

Ths patch uses the vb->isBackfillPhase() to determine if the vbucket is
in a backfill state as opposed to using an opencheckpointid of zero.

Change-Id: Ia977d6bf90e54fd1ceb8db4a9088b19d94d4bc8c
Reviewed-on: http://review.couchbase.org/70810
Reviewed-by: Manu Dhundi <manu@couchbase.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>