MB-21975: KV Error Map
The error map is a map of error codes to predefined error attributes
which allows clients to handle error codes they do not have explicit
knowledge of. The error attributes describe certain properties of each
error, and using these attributes, clients may determine the best
behavior.

This is intended to be used in conjunction with the `XERROR` feature:
Clients who advertise support for XERROR will need to get the error map
using the `PROTOCOL_BINARY_CMD_GET_ERROR_MAP` command, which will return
the error map in JSON format.

Change-Id: Ieb4a3a8ccb23522c5bd12d189e420c098510152d
Reviewed-on: http://review.couchbase.org/67175
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: buildbot <build@couchbase.com>