Refactor: Move xattr to its own library
The underlying engines must be able to get a basic knowledge
of xattrs in order to allow the multiple combinations of
DCP:

  * With and without value
  * With and without xattrs

The current implementation of DCP will strip off the body
before sending to memcached if no value is requested (and
that would also result in no xattrs).

It is desirable to allow ep-engine to avoid fetching data
off disk which would be discarded in memcached.

Change-Id: I26b352438bbbf3396a4aa99a8022781b9cd3395c
Reviewed-on: http://review.couchbase.org/71256
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>