Add workaround for memory leak in std::regex
There is a memory leak in std::regex on our valgrind commit
builders. This memory leak isn't something we care about as
we only create a single instance of it in the lifetime of
the program. Instead of having to modify the suppression
list passed to valgrind we can just use a unique pointer
in this case as we don't utilize the regexp when running
with valgrind.

Change-Id: If7c6bb46cb4f892e892f4836b3bd2430be204cf5
Reviewed-on: http://review.couchbase.org/71173
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: buildbot <build@couchbase.com>