Refactor: promote EWB protocol description
Promote the wire definition and opcode to protocol_binary.h.
This is where we would expect to locate the information, and
it makes it easier to avoid an opcode-crash when adding new
feature. At the same time I moved the opcode to 0xfd as
the previous value was located within a big range of
available opcodes.

Change-Id: I18ae6494a38db9c15dc9713d0a386eda3e2e9d73
Reviewed-on: http://review.couchbase.org/71270
Reviewed-by: Sriram Ganesan <sriram@couchbase.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>