Allow to build without support for Greenstack
Greenstack isn't currently enabled by the server, and
the project is currently put on hold while we're working
on other tasks. One alternative would be to completely
remove all of the support in our source directory,
but I think that the least invasive change (and the
easiest to bring back without conflicts) is just to
allow to disable the source.

Change-Id: Ic6d2b962d97ba96f9f92e12306098ab48d6e11d9
Reviewed-on: http://review.couchbase.org/71296
Reviewed-by: Daniel Owen <owend@couchbase.com>
Tested-by: buildbot <build@couchbase.com>