MB-21938: Shutdown memcached server cleanly in embedded mode
The memcached_testapp in embedded mode runs the memcached server in the
same process as the test.  When the testapp terminates the memcached
server thread is abrubtly terminated.

The result is that the Phosphor TraceLog destructor is called whilst
other memcached server threads are still making use of it.

The solution is to ensure that the memcached server is cleanly shutdown
and then performs a cb_join_thread to the main testapp thread.

The memcached server thread intialises SSL at start-up and cleans-up
during shutdown.  Therefore the McdEnvironment is changed to only
manage the SSL initalisation and shutdown when running in stand-alone
mode.

Change-Id: Ie854b5cb40529e924c080b38b72b8adef1e36bb1
Reviewed-on: http://review.couchbase.org/71476
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>