Refactor: Simplify inflate logic in couch-kvstore
Replace the explicit use of cb_malloc+cb_free with a
unique_ptr and use our cb::compression version over
the C api to snappy.

The old version of the code did not handle allocation
failures nor any problems with inflation. This
patch results in bad_alloc or runtime_error being
thrown. (which is probably better than using a
nullpointer or operating on uninitialized data)

Change-Id: I50719f4235192fa581c33f0960ca299df2b70229
Reviewed-on: http://review.couchbase.org/71488
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: buildbot <build@couchbase.com>