Remove VBucketVisitorTask and replace usage with VBCBAdaptor
The VBucketVisitorTask and VBCBAdaptor perform very similar tasks
and share most of their code. The primary difference that
the VBCBAdaptor has is it filters out the visited vbuckets to a
single shard. Given that the visitors they are given can filter out
vbuckets themselves (Using a VBucketFilter) it makes more sense
to do the filtering in one place.

This therefore updates the AccessScannerVisitor to use the
VBCBAdaptor instead of the VBucketVisitorTask and removes the
VBucketVisitorTask.

Change-Id: Iba1fd3995aacafc9ff7aa3333baab4815e001de2
Reviewed-on: http://review.couchbase.org/71497
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>