MB-22159: Use unsigned type for Task histograms
There is a potential out_of_range exception thrown, causing tests
attempting to parse the result to fail if a request is made for
runtimes or scheduler stats and a task has zero duration (or wait
time).

This should instead be changed to a unsigned type.

Note: This only affects tests, product should not be affected.

Change-Id: Ida99eb7acd5fb02384cdf6c3a945b05763ecd102
Reviewed-on: http://review.couchbase.org/71510
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>