Add unit test for password calculation
There is a suggestion that we should use the same storage
format for plain text passwords as ns_server do. By doing
that ns_server may add an entry for Administrator without
having to reset its password (as it doesn't know the
plain text version of it).

This change does _NOT_ change the way it is stored in
our files, just to make sure that we produce the same
result.

Change-Id: Ib4a0dc9408b1708f30e59cc5741493f2d99195d8
Reviewed-on: http://review.couchbase.org/71546
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>