Add callback function to map engine code to mcbp
Some commands in ep-engine tries to build up the response
message by converting an engine error code to mcbp status code.
That lookup table does not handle the new engine error codes,
nor does it handle if the client understands extended error
codes.

While we're fixing that we need to call back into memcached
and have it do the lookup for us.

Change-Id: Ic2f435a88efbc88af87dfd0651bc4dcd1d10c7fe
Reviewed-on: http://review.couchbase.org/71554
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>