MB-22082: Improve logging of connections marked as "want to disconnect"
Rebalance failure issues such as the one in the referenced MB are often
due to connections being forcably closed.  Furthermore these issues
can be hard to reproduce, and therefore it is important to be able to
debug from the logs.

Therefore this patch improves the logging for when we close a
connection by recording when we mark a connection as "want to
disconnect".

Change-Id: I122e685484557a96c36a61c16787bfcf0adf767f
Reviewed-on: http://review.couchbase.org/71559
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: buildbot <build@couchbase.com>