MB-22193: Allow lock and unlock through engine API
This removes the need to understand the mcbp protocol in
ep-engine for those two commands (and we may reuse the logic
for greenstack).

As an extra bonus this means that we can add unit tests
to operate on top of locked items (once we add support for
the operations in memcached buckets).

NOTE: This patch depend on a patch in ep-engine to move
the implementation in ep-engine over to this method.

Change-Id: Iab881dbf41a365e41370769affe1025a9756d6e9
Reviewed-on: http://review.couchbase.org/71599
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: buildbot <build@couchbase.com>