MB-22194: Fix data race on VBucketMemoryDeletionTask::vbucket
In VBucketMemoryDeletionTask::getDescription() the ref-counted pointer
`vbucket` can be made invalid (null) just after a check has been made
on it's correctness, if the ::run() method is executing in another
thread and it has just called vbucket.reset().

Fix to remove any possibility of concurrent access - calculate the
description once in the object constructor.

Change-Id: Ie2d3515a8b4f2dcfe5aec36e50e730a02aade462
Reviewed-on: http://review.couchbase.org/71600
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Manu Dhundi <manu@couchbase.com>