MB-22082: Improve logging when closing connections
Rebalance failure issues such as the one in the referenced MB are often
due to connections being forcably closed.  Furthermore these issues
can be hard to reproduce, and therefore it is important to be able to
debug from the logs.

Therefore this patch improves the logging for when we close a
connection because we have been unable to update the libevent settings
for a connection.

The log messages include both the connection id and the cookie.  The
cookie is included as it allows us to match-up connections logged by
ep-engine.

The log message for when we close a connection due to a socketError
is also updated to include the cookie.

Change-Id: Ibad7ecedb2f7c57a7418797c63525627871058f5
Reviewed-on: http://review.couchbase.org/71601
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>