Warmup: Don't create task when Warmup disabled
Do not create (or run) the Warmup task when ep_warmup is false.

This simplifies the on-lining of a bucket when Warmup is disabled - we
don't bother to run a bunch of background tasks which ultimately do
nothing (but do check various access log files, etc) - and means that
without Warmup buckets are immediately online after initialize()
finishes.

Adds some initial warmup unit tests - thus far just checking that
stats are reported correctly when warmup is disabled.

Change-Id: I1434dc3c2ccd77a6bb565a0df77c6ecc2d66b3a6
Reviewed-on: http://review.couchbase.org/71645
Reviewed-by: Daniel Owen <owend@couchbase.com>
Tested-by: buildbot <build@couchbase.com>